Fleet designations based on NEGEAR2 codes

NEGEAR2 codes are assigned to fleets based on which species are landed. The methodology and the decisions made to create this data set can be found below

fleets

Format

A data frame with 3 columns

fleetName

Name of fleet

NEGEAR2

Gear codes associated with each fleet

hydraFleets

Fleet aggregation to satify Hydra multispecies model

Methodology

This data set was created based on clustering methods.