The Atlantis-simulated survey age class specific length composition for the mskeyrun project

Simulated survey length composition data by age class from Atlanits model as sampled by atlantisom

Usage

simSurveyAgeLencomp

Format

A data frame

ModSim

Atlantis model name and simulation id

year

year simulated survey conducted

Code

Atlantis model three letter code for functional group

Name

Atlantis model common name for functional group

survey

simulated survey name

agecl

age class of Atlantis functional group

lenbin

1 cm length bin, lower limit

variable

number at length (Natlen)

value

value of the variable

units

units of the variable